:recycle: refactor: move model class into directory
