+++ index.html
@@ -15,7 +15,7 @@
         <form id="purchase-form">
           <label for="purchase-input">구입할 금액을 입력해주세요.</label>
           <input id="purchase-input" type="number" placeholder="금액" />
-          <button id="purchase-button" type="button">구입</button>
+          <button id="purchase-button" type="submit">구입</button>
         </form>
       </section>
       <section id="ticket-container">
