--- src/index.js
@@ -1,2 +1,0 @@
-import "./css/index";
-import "./js/app";
