+++ src/js/model/Lotto.js
